requirements.txt
fastapi~=0.115.12
numpy~=2.2.6
python-dotenv~=1.0.1
pydantic~=2.7.4
uvicorn~=0.32.1
sentence-transformers>=3.0.0
faiss-cpu
